Move SlotModal outside of settings folder
